+++ NEWS.md
@@ -1,3 +1,4 @@
 # Changes in v0.0.1
 
 + Initial package skeleton. 
++ Added main function `findVariableFeaturesBayes()` (still in development right now).